requirements.txt
Flask==2.1.1
Celery==5.1.2
redis==4.1.1
Pillow==8.4.0
pytesseract==0.3.8